requirements.txt
typer>=0.9
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
tenacity>=8.2
Pillow>=10.0
ebooklib>=0.18
readability-lxml>=0.8.1
pytest>=7.4


